Changed parameter type at all usage

+++ features/content/profile/src/main/java/com/tokopedia/profile/view/fragment/ProfileFragment.kt
@@ -1166,7 +1166,7 @@ class ProfileFragment : BaseListFragment<Visitable<*>, BaseAdapterTypeFactory>()
     override fun onGridItemClick(
         positionInFeed: Int,
         contentPosition: Int,
-        productPosition: Int,
+        productPosition: Long,
         redirectLink: String,
         type: String,
         isFollowed: Boolean,
