+++ features/merchant/shop_page/src/main/java/com/tokopedia/shop/feed/view/fragment/FeedShopFragment.kt
@@ -869,7 +869,7 @@ class FeedShopFragment : BaseListFragment<Visitable<*>, BaseAdapterTypeFactory>(
     override fun onGridItemClick(
         positionInFeed: Int,
         contentPosition: Int,
-        productPosition: Int,
+        productPosition: Long,
         redirectLink: String,
         type: String,
         isFollowed: Boolean,
